# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Booking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Booking:
      type: object
      properties:
        id:
          type: integer
        confirmationId:
          type: string
        service:
          type: string
        servicePlace:
          type: string
        date:
          type: string
        time:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        petName:
          type: string
        petBreed:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - confirmationId
        - service
        - servicePlace
        - date
        - time
        - ownerName
        - email
        - phone
        - petName
        - petBreed
        - notes
        - status
        - createdAt

    CreateBookingBody:
      type: object
      properties:
        service:
          type: string
        servicePlace:
          type: string
        date:
          type: string
        time:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        petName:
          type: string
        petBreed:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - service
        - servicePlace
        - date
        - time
        - ownerName
        - email
        - phone
        - petName
        - petBreed

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error